requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.3
django-crispy-forms==1.10.0
django-heroku==0.3.1
djangorestframework==3.12.2
factory-boy==3.1.0
Faker==5.0.2
gunicorn==20.0.4
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
static3==0.7.0
text-unidecode==1.3
waitress==1.4.4
whitenoise==5.2.0
